Update README.md
Add a code snippet to register installations. One of the main reasons why I (and others) use the AppsFlyer SDK is to track the success of ads placed on different platforms. The ad partners will always need the number of installations from the ads they have placed on your behalf. Had a hard time resolving issues with my ad partners as I only followed the instructions on the README.md file.